# Conflicts:
#	.idea/misc.xml
#	sql/drop.sql
#	src/main/java/at/ac/tuwien/sepm/ws16/qse01/gui/ShootingAdminController.java
#	src/test/java/at/ac/tuwien/sepm/ws16/qse01/dao/ShootingDAOTest.java

+++ src/main/java/at/ac/tuwien/sepm/ws16/qse01/gui/ShootingAdminController.java
@@ -102,7 +102,7 @@ public class ShootingAdminController {
                 try{
 
                     Profile profile = (Profile) profileChoiceBox.getSelectionModel().getSelectedItem();
-                    Shooting shouting = new Shooting((int) profile.getId(), path, true);
+                    Shooting shouting = new Shooting(profile.getId(), path, true);
 
                     storageDirLabel.setText("");
                     LOGGER.info("ShootingAdminController:", path);
